채팅방 유저입장에서 삭제 - 상대방은 삭제 안되게끔 구현

+++ src/main/java/com/example/giveandtake/service/ChatService.java
@@ -2,9 +2,11 @@ package com.example.giveandtake.service;
 
 import com.example.giveandtake.DTO.ChatMessageDTO;
 import com.example.giveandtake.DTO.ChatRoomDTO;
+import com.example.giveandtake.DTO.ReplyDTO;
 import com.example.giveandtake.common.CustomUserDetails;
 import com.example.giveandtake.model.entity.ChatMessage;
 import com.example.giveandtake.model.entity.ChatRoom;
+import com.example.giveandtake.model.entity.Reply;
 import com.example.giveandtake.repository.ChatMessageRepository;
 import com.example.giveandtake.repository.ChatRoomRepository;
 import lombok.AllArgsConstructor;
@@ -13,6 +15,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import java.nio.file.attribute.UserDefinedFileAttributeView;
 import java.security.Principal;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -49,16 +52,18 @@ public class ChatService {
         Collections.reverse(chatList);
         return chatList;
     }
+
     //특정채팅방 조회
     public List<ChatRoom> findRoomById(Long roomId) {
         return chatRoomRepository.findByRoomId(roomId);
     }
-   //대화내용 저장
+    //대화내용 저장
     public void createMessage(ChatMessageDTO chatMessageDTO) {
 
         if (ChatMessage.MessageType.QUIT.equals(chatMessageDTO.getType())) {
             chatMessageDTO.setMessage(chatMessageDTO.getSender() + "님이 방에서 나갔습니다.");
             chatMessageDTO.setSender("[알림]");
+
         }
         Long msgNum = chatMessageRepository.save(chatMessageDTO.toEntity()).getMsgNum();
         Optional <ChatMessage> chatMessageList = chatMessageRepository.findByMsgNum(msgNum);
@@ -66,16 +71,30 @@ public class ChatService {
 
         messagingTemplate.convertAndSend("/sub/chat/room/" + chatMessageDTO.getRoomId(), chatMessage);
 
-
     }
     //채팅방 삭제
     public void deleteChatRoom(Long roomId, Principal principal) {
+        ChatRoomDTO chatRoomDTO =new ChatRoomDTO();
+        List<ChatRoom> chatRoom = chatRoomRepository.findByRoomId(roomId);
+        ChatRoom chats = chatRoom.get(0);
 
-        System.out.println("%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%끝 delete"+roomId);
-        
+        String name = principal.getName();
+        chatRoomDTO.setRoomId(roomId);
+        chatRoomDTO.setRoomName(chats.getRoomName());
+        chatRoomDTO.setReceiver(chats.getReceiver());
+        chatRoomDTO.setRequest(chats.getRequest());
 
-        chatRoomRepository.deleteById(roomId);
-
+        if( name.equals(chats.getReceiver()) ){
+            chatRoomDTO.setReceiver("");
+            chatRoomRepository.save(chatRoomDTO.toEntity()).getRoomId();
+        }
+        if(name.equals(chats.getRequest())){
+            chatRoomDTO.setRequest("");
+            chatRoomRepository.save(chatRoomDTO.toEntity()).getRoomId();
+        }
+        if(chats.getRequest().equals("") && chats.getReceiver().equals("")) {
+            chatRoomRepository.deleteById(roomId);
+        }
     }
 
     //채팅방 메시지 리스트 가져오기
